+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Dockerised development setup, with support for live reload of `flowmachine` and `flowapi` after source code changes.
 - Pre-commit hook for Python formatting with black.
-- Added new `IntereventPeriod` query classes to FlowMachine.
+- Added new `IntereventPeriod`, `ContactReciprocal`, `ProportionContactReciprocal`, `ProportionEventReciprocal`, `ProportionEventType`, `MDSVolume` query classes to FlowMachine.
 
 ### Changed
 - `CustomQuery` now requires column names to be specified
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Query objects can no longer be recalculated to cache and must be explicitly removed first
 - Arbitrary `Flow` maths
 - `EdgeList` query type
+- Removes query class `ProportionOutgoing` as it becomes redundant with the the introduction of `ProportionEventType`.
 
 
 ## [0.3.0]